(server) unit test 1
